pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hcg-kg"
version = "0.1.0"
description = "Source-grounded biomedical knowledge graph construction and query pipeline for parsed clinical guideline JSON."
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.11"
authors = [
  { name = "HeartBioPortal contributors" }
]
keywords = ["biomedical-nlp", "knowledge-graph", "clinical-guidelines", "neo4j", "heartbioportal"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Science/Research",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: Apache Software License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
  "Topic :: Scientific/Engineering :: Bio-Informatics",
  "Topic :: Software Development :: Libraries :: Python Modules"
]
dependencies = [
  "joblib>=1.4.2",
  "networkx>=3.3",
  "numpy>=1.26",
  "pydantic>=2.8",
  "pyyaml>=6.0.2",
  "rich>=13.9",
  "scikit-learn>=1.5.1",
  "typer>=0.15.1",
]

[project.optional-dependencies]
dev = [
  "mypy>=1.11",
  "pre-commit>=3.8",
  "pytest>=8.3",
]
neo4j = [
  "neo4j>=5.24",
]
llm = [
  "llama-index>=0.12",
  "ollama>=0.4",
  "sentence-transformers>=3.0",
  "transformers>=4.44",
]

[project.scripts]
hcg-kg = "hcg_kg.cli:app"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
  "E",
  "F",
  "I",
  "UP",
  "B",
  "SIM",
  "C4",
]

[tool.ruff.format]
quote-style = "double"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
disallow_any_generics = false
plugins = []

